Move methods from SecuredFunctions to ReadJdbcP [HZ-3427]
